requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aniso8601==7.0.0
apispec==3.3.0
apistar==0.7.2
APScheduler==3.6.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
discord.py==1.3.1
docopt==0.6.2
Flask==1.1.1
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
h11==0.9.0
httptools==0.0.13
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.0
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.4
packaging==20.1
pluggy==0.13.1
promise==2.3
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-multipart==0.0.5
python-twitch-client==0.6.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-toolbelt==0.9.1
responder==2.0.5
rfc3986==1.3.2
Rx==1.6.1
six==1.14.0
SQLAlchemy==1.3.13
starlette==0.12.13
typesystem==0.2.4
tzlocal==2.0.0
urllib3==1.25.8
uvicorn==0.10.9
uvloop==0.14.0
wcwidth==0.1.8
websockets==8.1
Werkzeug==1.0.0
whitenoise==5.0.1
wrapt==1.11.2
yarl==1.4.2
